Implementation of integration test for subtensor v1.1.0 [Hold review, needs more work] (#286)
* Fixed bugs in both subtensor client as well as the corresponding integration test

* Moved subtensor integration test to ./tests/integration_tests/subtensor/v1.0.1/client.py

* Added WalletStub

* Reimplemented removed tests

* Updated circle ci config to run the subtensor integration tests

* Changed dockerfile

* Added client test for subtensor v1.1.0

* Removed duplicate tests

* Removed duplicate tests

* Fixed add_stake and unstake tests for v1.1.0

* Added set_weights test for v1.1.0 that checks transaction fee

* Finished tranfer test to include a check for transaction fees

* Finalized integration tests for subtensor v1.1.0

* Fixed issue in dockerf
Updated config to pull different image

* Hope this fixes

* Updated dockerfile

* Fixed bug in push_image script

* Added downloading of release v1.1.0 to docker

* Enabled execution of subtensor v1.1.0 integration tests

* Commented out unit tests of subtensor, as these are actually integration tests, which have been implmented in the integrate_tests folder.

They are commented out instead of deleted, as they may provide useful if some additional mocking is done

* Removed weird line

* Fixed subtensor integration tests to work with modality

* Switch docker image for circleci over to latest
