+++ .circleci/config.yml
@@ -89,6 +89,25 @@ jobs:
           name: "Code Coverage"
           command: codecov
 
+  "Run subtensor integration tests":
+    executor: bittensor-executor
+    steps:
+      - checkout
+      - run: pip3 install cython
+      - run: pip3 install -r requirements.txt
+      - run: ./scripts/build_protos.sh
+      - run: ./scripts/create_wallet.sh
+      - run:
+          command: pip3 install -e .
+          name: setup
+      - run:
+          name: "Run integration tests for subtensor v1.0.1"
+          command: "export PYTHONUNBUFFERED=1; export NODE_SUBTENSOR_BIN=/subtensor/v1.0.1/node-subtensor; pytest -n 8 ./tests/integration_tests/subtensor/v1.0.1/*"
+      - run:
+          name: "Run integration tests for subtensor v1.1.0"
+          command: "export PYTHONUNBUFFERED=1; export NODE_SUBTENSOR_BIN=/subtensor/v1.1.0/node-subtensor; pytest -n 8 ./tests/integration_tests/subtensor/v1.1.0/*"
+
+
   "Push image to DockerHub":
     docker:
       - image: circleci/python:3.8.1
@@ -118,6 +137,9 @@ workflows:
               requires:
                 - "Build Bittensor"
       - "Run Integration Tests":
+              requires:
+                - "Build Bittensor"
+      - "Run subtensor integration tests":
               requires:
                 - "Build Bittensor"
       - "Push image to DockerHub":
