+++ scripts/build_protos.sh
@@ -1,1 +1,1 @@
-python3.7 -m grpc.tools.protoc bittensor/bittensor.proto  -I. --python_out=. --grpc_python_out=.
+python3 -m grpc.tools.protoc bittensor/bittensor.proto  -I. --python_out=. --grpc_python_out=.
