+++ scripts/push_image.sh
@@ -1,3 +1,3 @@
 ./scripts/build_protos.sh
-docker build -t bittensor/bittensor:latest -f Dockerfile .
-docker push bittensor/bittensor:latest
+docker build -t bittensor/bittensor:subtensor_tests -f Dockerfile .
+docker push bittensor/bittensor:subtensor_tests
